+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/Payload.java
@@ -95,7 +95,8 @@ public abstract class Payload {
         INIT_CONTRACT,
         DEPLOY_MODULE,
         UPDATE,
-        TRANSFER_TO_PUBLIC
+        TRANSFER_TO_PUBLIC,
+        TRANSFER_TO_ENCRYPTED,
     }
 
 }